TUTB_artxiboa/TUTB_artxiboa/tutb_backend/requirements.txt
flask>=3.0
flask-cors>=4.0
mysql-connector-python>=8.0
PyJWT>=2.8
werkzeug>=3.0
